Preserve the deserialized value.
